squash: implement sortByIndex & move sort options into sorting library module
